feat: migrate ProfileScreen to :feature:profile KMP module
- Create :feature:profile KMP module with androidTarget + iOS targets
- Move ProfileScreenStateless + ProfileUiState to commonMain
- App keeps stateful wrapper with koinViewModel and previews
